Version 1.0
This was the first sloppy version of the bus tracker. I just wanted to
see if I could do it and get it live. It works just fine but it is
spaghetti code and not efficient. The only benefit of this one over the
later versions is that it can take more than 2 locations a day.