Version 3.0
This version is an overhaul of the previous. It acts similar in the
logic but much of the code is changed. Jquery is now used as it is much
easier to make changes to the DOM when needed. The schedule is no longer
a huge array of strings and is now on a database on the website server.
It uses AJAX to access a file that the server is setup to generate which
has the most up to date schedules.